Crate db_sync_follower

Provides implementations of Data Sources that read from db-sync postgres

Modules

block

candidates

data_sources

Data sources implementations that read from db-sync postgres.

mc_hash

metrics

native_token

sidechain_rpc

Macros

observed_async_trait

Logs each method invocation and each returned result. Has to be made at the level of trait, because otherwise #async_trait is expanded first. ‘&self’ matching yields “__self” identifier not found error, so “&$self:tt” is required. Works only if return type is Result.

Structs

SqlxError

Enums

DataSourceError

Type Aliases

Result